requirements.txt
biopython
streamlit
pandas
matplotlib